api/api_cat_dog/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
aiofiles==23.2.1